+++ src/test/java/com/expedia/seisoimport/controllers/BuildVersionSchedulerTest.java
@@ -4,6 +4,6 @@ package com.expedia.seisoimport.controllers;
  * Author: James McQueen (jmcqueen@expedia.com)
  * Created: 1/25/16
  */
-public class MessageRecieverTest
+public class BuildVersionSchedulerTest
 {
 }
