+++ src/test/java/com/expedia/seisoimport/retrievers/SQSRetrieverTest.java
@@ -1,9 +1,9 @@
-package com.expedia.seisoimport.services;
+package com.expedia.seisoimport.retrievers;
 
 /**
  * Author: James McQueen (jmcqueen@expedia.com)
  * Created: 3/6/16
  */
-public class BuildVersionServiceTest
+public class SQSRetrieverTest
 {
 }
